backend/requirements.txt
Django
djangorestframework
pandas
openpyxl
django-cors-headers
